requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file='d:\projects\CSA\csa-hello\requirements.txt' 'd:\projects\CSA\csa-hello\requirements.in'
#
alembic==1.15.2
    # via -r d:\projects\CSA\csa-hello\requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
asyncpg==0.30.0
    # via -r d:\projects\CSA\csa-hello\requirements.in
cffi==1.17.1
    # via cryptography
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via click
cryptography==44.0.2
    # via
    #   jwt
    #   python-jose
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via pydantic
fastapi==0.115.12
    # via -r d:\projects\CSA\csa-hello\requirements.in
fpdf==1.7.2
    # via -r d:\projects\CSA\csa-hello\requirements.in
greenlet==3.2.1
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
jinja2==3.1.6
    # via -r d:\projects\CSA\csa-hello\requirements.in
jwt==1.3.1
    # via -r d:\projects\CSA\csa-hello\requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
passlib==1.7.4
    # via -r d:\projects\CSA\csa-hello\requirements.in
psycopg2-binary==2.9.10
    # via -r d:\projects\CSA\csa-hello\requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pycryptodome==3.22.0
    # via -r d:\projects\CSA\csa-hello\requirements.in
pydantic[email]==2.11.4
    # via
    #   -r d:\projects\CSA\csa-hello\requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.0
    # via -r d:\projects\CSA\csa-hello\requirements.in
python-jose[cryptography]==3.4.0
    # via -r d:\projects\CSA\csa-hello\requirements.in
python-multipart==0.0.20
    # via -r d:\projects\CSA\csa-hello\requirements.in
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.40
    # via
    #   -r d:\projects\CSA\csa-hello\requirements.in
    #   alembic
starlette==0.46.2
    # via fastapi
typing-extensions==4.13.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
uvicorn==0.34.2
    # via -r d:\projects\CSA\csa-hello\requirements.in
websockets==15.0.1
    # via -r d:\projects\CSA\csa-hello\requirements.in
xlsxwriter==3.2.3
    # via -r d:\projects\CSA\csa-hello\requirements.in
